Commited the whole code
Commited the whole code together with the pstreams.h source file from (http://pstreams.sourceforge.net/)  that is licensed under Boost Software License that is included in the same directory as the pstreams.h file.

My code is licensed under the Boost Software License as well. So, everything in this repository is under said license. 

The code heavily relies on sfml library and to make it compile one has to download sfml first.